requirements.txt
numpy
scipy
matplotlib
tqdm
snapgene_reader
numba
icecream
genomepy
biopython